test(): move recommendedTemplatesCatalog to own module
