+++ packages/common/src/initiative-templates/_internal/InitiativeTemplateUiSchemaEdit.ts
@@ -1,12 +1,7 @@
-import { IUser } from "@esri/arcgis-rest-types";
 import { IArcGISContext } from "../..";
 import { IHubInitiativeTemplate } from "../../core";
 import { IUiSchema, UiSchemaMessageTypes } from "../../core/schemas/types";
-import { IHubCollection } from "../../search";
-import {
-  getWellKnownCatalog,
-  WellKnownCatalog,
-} from "../../search/wellKnownCatalog";
+import { getRecommendedTemplatesCatalog } from "./getRecommendedTemplatesCatalog";
 
 /**
  * @private
@@ -115,91 +110,10 @@ export const buildUiSchema = async (
               operation: "OR",
             },
           ],
-          drag: false,
+          canReorder: false,
           linkTarget: "workspaceRelative",
         },
       },
     ],
-  };
-};
-
-const getRecommendedTemplatesCatalog = (user: IUser, i18nScope: string) => {
-  const catalogNames: WellKnownCatalog[] = [
-    "myContent",
-    "favorites",
-    "organization",
-    "world",
-  ];
-
-  const catalogs = catalogNames.map((name: WellKnownCatalog) => {
-    const opts = { user };
-    const catalog = getWellKnownCatalog(
-      "initiativeTemplate.fields.recommendedTemplates",
-      name,
-      "item",
-      opts
-    );
-
-    // manually attach recommended templates collection
-    catalog.collections = [getRecommendedTemplatesCollection(i18nScope)];
-    return catalog;
-  });
-
-  return catalogs;
-};
-
-const getRecommendedTemplatesCollection = (
-  i18nScope: string
-): IHubCollection => {
-  return {
-    targetEntity: "item",
-    key: "recommendedTemplates",
-    label: `${i18nScope}.fields.recommendedTemplates.collection.label`,
-    scope: {
-      targetEntity: "item",
-      filters: [
-        {
-          predicates: [
-            {
-              typekeywords: {
-                not: ["hubSolutionType|hubSiteApplication"],
-              },
-            },
-          ],
-        },
-        {
-          predicates: [
-            {
-              typekeywords: {
-                any: [
-                  "hubSolutionType|storymap",
-                  "hubSolutionType|webmap",
-                  "hubSolutionType|dashboard",
-                  "hubSolutionType|hubpage",
-                  "hubSolutionType|webexperience",
-                  "hubSolutionType|webmappingapplication",
-                  "hubSolutionType|form",
-                  "hubSolutionType|featureservice",
-                  "Template",
-                ],
-              },
-            },
-          ],
-          operation: "OR",
-        },
-        {
-          predicates: [
-            {
-              typekeywords: ["hubSolutionTemplate"],
-            },
-            {
-              type: "Solution",
-              typekeywords: ["Template"],
-            },
-          ],
-          operation: "OR",
-        },
-      ],
-    },
   };
 };
